🔨 organised cv
