+++ client/src/Curriculum.js
@@ -5,15 +5,15 @@ import "./css/global.css";
 
 
 import NavigationBar from "./components/navigationBar/NavigationBar";
-import CardSection from "./components/cv/CardSection";
+import Cv from "./components/cv/Cv";
 import Footer from "./components/footer/Footer";
 
-export default function Cv(props) {
+export default function Curriculum(props) {
 
   return (
     <div className="wideGrid">
       <NavigationBar {...props} />
-      <CardSection {...props} />
+      <Cv {...props} />
       <Footer />
     </div>
   );
